+++ src/main/java/com/example/naengtal/domain/member/service/AccountService.java
@@ -1,7 +1,7 @@
 package com.example.naengtal.domain.member.service;
 
 import com.example.naengtal.domain.fridge.entity.Fridge;
-import com.example.naengtal.domain.fridge.repository.FridgeRepository;
+import com.example.naengtal.domain.fridge.dao.FridgeRepository;
 import com.example.naengtal.domain.member.dao.MemberRepository;
 import com.example.naengtal.domain.member.dto.MemberInfo;
 import com.example.naengtal.domain.member.dto.SignUpRequestDto;
